Fix the definition of union semun so that it is valid on systems where sizeof(long) != sizeof(int).  This resolves a hang on HP-UX/Itanium.
Submitted by:	Madhusudan Mathihalli <madhusudan_mathihalli@hp.com>
Reviewed by:	Jeff Trawick


git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@63526 13f79535-47bb-0310-9956-ffa450edef68

+++ include/arch/unix/proc_mutex.h
@@ -142,7 +142,7 @@ extern const apr_proc_mutex_unix_lock_methods_t apr_proc_mutex_unix_rwlock_metho
 
 #if !APR_HAVE_UNION_SEMUN && defined(APR_HAS_SYSVSEM_SERIALIZE)
 union semun {
-    long val;
+    int val;
     struct semid_ds *buf;
     unsigned short *array;
 };
